// server/src/api/V1/enums/Events.ts
enum events {
  JOIN_GAME = "JOIN GAME",
  UPDATE_GAME = "UPDATE GAME",
  MOVE = "MOVE",
  NEW_GAME = "NEW GAME",
  RESIGNATION = "RESIGNATION",
  DRAW_OFFER = "DREAW OFFER",
  DRAW_REJECT = "DRAW REJECT",
  DRAW_ACCEPT = "DRAW ACCEPT",
  CHAT_MESSAGE = "CHAT MESSAGE",
  DISCONECT = "disconnect",
}

export default events;